Move UserLayer stuff from control-base and service-map to control-userlayer and service-userlayer
