+++ control-base/pom.xml
@@ -56,7 +56,10 @@
             <groupId>org.oskari</groupId>
             <artifactId>service-print</artifactId>
         </dependency>
-
+        <dependency>
+            <groupId>org.oskari</groupId>
+            <artifactId>service-userlayer</artifactId>
+        </dependency>
         <dependency>
             <groupId>fi.nls.oskari</groupId>
             <artifactId>shared-test-resources</artifactId>
