+++ control-base/src/main/java/fi/nls/oskari/control/view/PublishPermissionHelper.java
@@ -20,7 +20,6 @@ import fi.nls.oskari.map.analysis.domain.AnalysisLayer;
 import fi.nls.oskari.map.analysis.service.AnalysisDbService;
 import fi.nls.oskari.map.analysis.service.AnalysisDbServiceMybatisImpl;
 import fi.nls.oskari.map.layer.OskariLayerService;
-import fi.nls.oskari.map.userlayer.service.UserLayerDbService;
 import fi.nls.oskari.myplaces.MyPlacesService;
 import fi.nls.oskari.permission.domain.Permission;
 import fi.nls.oskari.permission.domain.Resource;
@@ -30,6 +29,7 @@ import fi.nls.oskari.util.ConversionHelper;
 
 import org.json.JSONArray;
 import org.json.JSONObject;
+import org.oskari.map.userlayer.service.UserLayerDbService;
 import org.oskari.service.util.ServiceFactory;
 
 import java.util.ArrayList;
