+++ control-userlayer/src/main/resources/META-INF/services/javax.annotation.processing.Processor
@@ -1,0 +1,5 @@
+# This Annotation processor will be run automatically by the compiler
+# in order to detect all of the places that the @OskariActionRoute annotation
+# is used.
+
+fi.nls.oskari.annotation.OskariActionRouteAnnotationProcessor
