+++ service-userlayer/src/main/java/org/oskari/map/userlayer/service/UserLayerDataService.java
@@ -1,4 +1,4 @@
-package fi.nls.oskari.map.userlayer.service;
+package org.oskari.map.userlayer.service;
 
 import fi.nls.oskari.domain.User;
 import fi.nls.oskari.domain.map.OskariLayer;
