+++ service-userlayer/src/main/java/org/oskari/map/userlayer/service/UserLayerDbService.java
@@ -1,4 +1,4 @@
-package fi.nls.oskari.map.userlayer.service;
+package org.oskari.map.userlayer.service;
 
 import fi.nls.oskari.domain.map.userlayer.UserLayer;
 import fi.nls.oskari.domain.map.userlayer.UserLayerData;
